+++ src/fr/maxlego08/menu/command/commands/CommandMenuVersion.java
@@ -18,6 +18,7 @@ public class CommandMenuVersion extends VCommand {
 
         message(sender, "§aVersion du plugin§7: §2" + plugin.getDescription().getVersion());
         message(sender, "§aAuteur§7: §2Maxlego08");
+        message(sender, "§aMarketplace/Inventory builder§7: §2https://minecraft-inventory-builder.com/");
         message(sender, "§aDiscord§7: §2http://discord.groupez.dev/");
         message(sender, "§aDownload here§7: §2https://groupez.dev/resources/253");
         message(sender, "§aSponsor§7: §chttps://serveur-minecraft-vote.fr/?ref=345");