+++ changelog.md
@@ -33,6 +33,8 @@
 
 # Unreleased
 
+- Add custom sound to SoundOption
+
 # 1.0.2.4
 
 - Added replacement of , by . for  placeholders requirements. If your placeholder returns a number with a comma instead of a period, the plugin can handle that.